feat: 마이페이지와 거래지역 관리 사이드바 통합
               - 마이페이지에서 거래지역 관리 사이드바 연동
               - TradingAreaContent 컴포넌트 분리
               - useSidebar 훅으로 사이드바 열기 구현
               - 키보드 네비게이션 포커스 이슈 해결
               - 구매/판매 상품 더미데이터 추가
               - 상품 그리드 중앙 정렬 UI 개선
