+++ src/app/(user)/mypage/mypage.css
@@ -1,4 +1,3 @@
-/* MyPage Component Styles */
 .mypage-container {
     max-width: 1440px;
     width: 1091px;
@@ -344,6 +343,35 @@
     line-height: 1.5;
     color: #666666;
     text-align: center;
+}
+
+/* 상품 목록 그리드 */
+.tab-content-area {
+    width: 100%;
+    padding: 20px 0;
+}
+
+.item-count {
+    font-size: 16px;
+    font-weight: 600;
+    color: #333;
+    margin-bottom: 16px;
+}
+
+.products-grid {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 16px;
+    width: 100%;
+    justify-content: flex-start;
+    max-width: calc(5 * 157px + 4 * 16px); /* 5개 카드 + 간격 */
+    margin: 0 auto; /* 전체 그리드를 중앙 정렬 */
+}
+
+.empty-state {
+    text-align: center;
+    color: #999;
+    padding: 40px 0;
 }
 
 /* Responsive Design */
